+++ event-details/arrival-tracking/index.ts
@@ -4,3 +4,5 @@ export * from "./Refresh"
 export * from "./ArrivalBanner"
 export * from "./ArrivalsOperation"
 export * from "./Models"
+export * from "./region-monitoring"
+export * from "./geofencing"
